Create simple square grid
A container div has been added to index page to store boxes.

javascript.js creates a grid of gridSize amount of boxes and stores in
both array for easy iteration but also to DOM in container.

CSS styles.css creates the visual grid using a flexbox on container,
and adds colors and borders to better display grid.